Refactored acces to featuresources/datastores/utility functions using datastores
