+++ solr-commons/src/main/java/nl/b3p/viewer/solr/SolrUpdateJob.java
@@ -16,6 +16,7 @@
  */
 package nl.b3p.viewer.solr;
 
+import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.Point;
@@ -135,7 +136,7 @@ public class SolrUpdateJob implements Job {
             List<String> resultAttributesConfig = config.getResultAttributes();
             
             SimpleFeatureType sft = config.getSimpleFeatureType();
-            fs = sft.openGeoToolsFeatureSource();
+            fs = SourceFactoryHelper.openGeoToolsFeatureSource(sft);
 
             Query q = new Query();
             if(filter != null){
