+++ viewer-admin/src/main/java/nl/b3p/viewer/admin/stripes/AttributeSourceActionBean.java
@@ -26,6 +26,10 @@ import net.sourceforge.stripes.validation.*;
 import nl.b3p.i18n.LocalizableActionBean;
 import nl.b3p.viewer.config.security.Group;
 import nl.b3p.viewer.config.services.*;
+import nl.b3p.viewer.helpers.featuresources.FeatureSourceHelper;
+import nl.b3p.viewer.helpers.featuresources.JDBCSourceHelper;
+import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
+import nl.b3p.viewer.helpers.featuresources.WFSSourceHelper;
 import nl.b3p.viewer.solr.SolrInitializer;
 import nl.b3p.web.WaitPageStatus;
 import org.apache.commons.logging.Log;
@@ -181,7 +185,7 @@ public class AttributeSourceActionBean extends LocalizableActionBean {
 
             JDBCFeatureSource fs = new JDBCFeatureSource(params);
             fs.setName(name);
-            fs.loadFeatureTypes(status);
+            JDBCSourceHelper.loadFeatureTypes(fs, status);
 
             em.persist(fs);
             em.getTransaction().commit();
@@ -196,7 +200,7 @@ public class AttributeSourceActionBean extends LocalizableActionBean {
             params.put(WFSDataStoreFactory.PASSWORD.key, password);
             WFSFeatureSource fs = new WFSFeatureSource(params);
             fs.setName(name);
-            fs.loadFeatureTypes(status);
+            WFSSourceHelper.loadFeatureTypes(fs, status);
             em.persist(fs);
             em.getTransaction().commit();
