+++ viewer-config-persistence-helpers/src/main/java/nl/b3p/viewer/helpers/featuresources/SourceHelper.java
@@ -1,0 +1,17 @@
+package nl.b3p.viewer.helpers.featuresources;
+
+import nl.b3p.viewer.config.services.FeatureSource;
+import nl.b3p.viewer.config.services.SimpleFeatureType;
+import nl.b3p.web.WaitPageStatus;
+import org.geotools.feature.FeatureCollection;
+import org.opengis.filter.Filter;
+
+import java.util.List;
+
+public interface SourceHelper {
+    List<SimpleFeatureType> createFeatureTypes(FeatureSource fs) throws Exception;
+    List<SimpleFeatureType> createFeatureTypes(FeatureSource fs, WaitPageStatus status) throws Exception;
+
+    org.geotools.data.FeatureSource openGeoToolsFeatureSource(FeatureSource fs, SimpleFeatureType sft) throws Exception;
+    org.geotools.data.FeatureSource openGeoToolsFeatureSource(FeatureSource fs, SimpleFeatureType sft, int timeout) throws Exception;
+}
