+++ viewer-config-persistence-helpers/src/main/java/nl/b3p/viewer/helpers/services/GeoserviceHelper.java
@@ -1,25 +1,18 @@
-package nl.b3p.viewer;
+package nl.b3p.viewer.helpers.services;
 
 import nl.b3p.viewer.config.ClobElement;
 import nl.b3p.viewer.config.services.GeoService;
 import nl.b3p.viewer.config.services.Layer;
-import nl.b3p.web.WaitPageStatus;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.mutable.MutableObject;
 import org.stripesstuff.stripersist.Stripersist;
 
 import javax.persistence.EntityManager;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 
 public class GeoserviceHelper {
-
-
-
-
 
     public static void initLayerCollectionsForUpdate(GeoService service) {
         EntityManager em = Stripersist.getEntityManager();
@@ -32,9 +25,7 @@ public class GeoserviceHelper {
         em.createQuery("from Layer l left join fetch l.children where l.service = :this").setParameter("this", service).getResultList();
     }
 
-
     public static void setAllChildrenDetail(Layer layer, EntityManager em) {
-
         layer.accept(new Layer.Visitor() {
 
             @Override
