+++ viewer-config-persistence/src/main/java/nl/b3p/viewer/config/services/FeatureSource.java
@@ -124,121 +124,6 @@ public abstract class FeatureSource {
     public String getProtocol() {
         return getClass().getAnnotation(DiscriminatorValue.class).value();
     }    
-    
-    List<String> calculateUniqueValues(SimpleFeatureType sft, String attributeName, int maxFeatures, Filter filter) throws Exception{
-        org.geotools.data.FeatureSource fs = null;
-        try {
-            FilterFactory2 ff = CommonFactoryFinder.getFilterFactory2(null);
-            Function unique = ff.function("Collection_Unique", ff.property(attributeName));
-            Filter notNull = ff.not( ff.isNull( ff.property(attributeName) ));
-            Filter f = notNull;
-            if(filter != null){
-                f = ff.and(notNull, filter);
-            }
-            org.geotools.data.Query q = new org.geotools.data.Query(sft.getTypeName(), f);
-            if(maxFeatures != -1) {
-                q.setMaxFeatures(maxFeatures);
-            }
-        
-            fs = sft.openGeoToolsFeatureSource();
-            FeatureCollection fc = fs.getFeatures(q);
-            
-            Object o = unique.evaluate( fc);
-            Set<String> uniqueValues  = (Set<String>)o;
-            if(uniqueValues == null){
-                uniqueValues = new HashSet<String>();
-            }
-            List<String> l = new ArrayList<String>(uniqueValues);
-            Collections.sort(l);
-            return l;
-        } catch (Exception ex) {
-            throw ex;
-        }finally{
-            if(fs != null && fs.getDataStore() != null){
-                fs.getDataStore().dispose();
-            }
-        }
-    }
-
-    public Map<String, String> getKeyValuePairs(SimpleFeatureType sft, String key, String label, int maxFeatures) throws Exception {
-        Map<String, String> output = new TreeMap< String, String>();
-        SimpleFeatureSource fs = null;
-
-        try {
-            fs = (SimpleFeatureSource) sft.openGeoToolsFeatureSource();
-
-            FilterFactory2 ff = CommonFactoryFinder.getFilterFactory2(null);
-            Filter notNull = ff.not(ff.isNull(ff.property(key)));
-            org.geotools.data.Query q = new org.geotools.data.Query(sft.getTypeName(), notNull);
-            q.setMaxFeatures(maxFeatures);
-            q.setPropertyNames(new String[]{key, label});
-        
-            SimpleFeatureIterator iterator = fs.getFeatures(q).features();
-            try {
-                while (iterator.hasNext()) {
-                    SimpleFeature f = iterator.next();
-                    output.put(
-                            f.getAttribute(key).toString(),
-                            f.getAttribute(label).toString()
-                    );
-                }
-            } finally {
-                iterator.close();
-            }
-            return output;
-        } catch (Exception ex) {
-            throw ex;
-        } finally {
-            if (fs != null && fs.getDataStore() != null) {
-                fs.getDataStore().dispose();
-            }
-        }
-    }
-
-    public Object getMaxValue(SimpleFeatureType sft, String attributeName, int maxFeatures, Filter f) throws Exception {
-        org.geotools.data.FeatureSource fs = null;
-        try {
-            FilterFactory2 ff = CommonFactoryFinder.getFilterFactory2(null);
-            Function max = ff.function("Collection_Max", ff.property(attributeName));
-
-            fs = sft.openGeoToolsFeatureSource();
-            FeatureCollection fc = f != null ? fs.getFeatures(f) : fs.getFeatures();
-            Object value = max.evaluate(fc);
-            return value;
-        } catch (Exception ex) {
-            throw ex;
-        } finally {
-            if (fs != null && fs.getDataStore() != null) {
-                fs.getDataStore().dispose();
-            }
-        }
-    }
-
-    public Object getMinValue(SimpleFeatureType sft, String attributeName, int maxFeatures, Filter filter) throws Exception {
-        org.geotools.data.FeatureSource fs = null;
-        try {
-            FilterFactory2 ff = CommonFactoryFinder.getFilterFactory2(null);
-            Function minFunction = ff.function("Collection_Min", ff.property(attributeName));
-            fs = sft.openGeoToolsFeatureSource();
-            
-            FeatureCollection f = filter != null ? fs.getFeatures(filter) : fs.getFeatures();
-
-            Object o = minFunction.evaluate(f);
-            return o;
-        } catch (Exception ex) {
-            throw ex;
-        } finally {
-            if (fs != null && fs.getDataStore() != null) {
-                fs.getDataStore().dispose();
-            }
-        }
-    }
-    
-    /* package */ abstract org.geotools.data.FeatureSource openGeoToolsFeatureSource(SimpleFeatureType sft) throws Exception;
-    
-    /* package */ abstract FeatureCollection getFeatures(SimpleFeatureType sft, Filter f, int maxFeatures) throws Exception;
-    
-    /* package */ abstract org.geotools.data.FeatureSource openGeoToolsFeatureSource(SimpleFeatureType sft, int timeout) throws Exception;
 
     public SimpleFeatureType getFeatureType(String typeName) {
         for(SimpleFeatureType sft: getFeatureTypes()) {
