+++ viewer-config-persistence/src/main/java/nl/b3p/viewer/config/services/JDBCFeatureSource.java
@@ -40,13 +40,12 @@ import org.opengis.filter.Filter;
  */
 @Entity
 @DiscriminatorValue(JDBCFeatureSource.PROTOCOL)
-public class JDBCFeatureSource extends UpdatableFeatureSource{
-    private static final Log log = LogFactory.getLog(JDBCFeatureSource.class);
+public class JDBCFeatureSource extends FeatureSource{
 
     public static final String PROTOCOL = "jdbc";
 
     @Column(name="db_schema")
-    private String schema;
+    public String schema;
 
     public String getSchema() {
         return schema;
@@ -73,180 +72,5 @@ public class JDBCFeatureSource extends UpdatableFeatureSource{
         schema = (String)params.get("schema");
         setUsername((String)params.get("user"));
         setPassword((String)params.get("passwd"));
-    }
-
-    public void loadFeatureTypes() throws Exception {
-        loadFeatureTypes(new WaitPageStatus());
-    }
-
-    public void loadFeatureTypes(WaitPageStatus status) throws Exception {
-        this.getFeatureTypes().addAll(createFeatureTypes(status));
-    }
-
-    /**
-     * Creates list of featuretypes for this FeatureSource.
-     *
-     * @return list of featuretypes
-     * @throws java.lang.Exception if any
-     */
-    public List<SimpleFeatureType> createFeatureTypes() throws Exception{
-        return createFeatureTypes(new WaitPageStatus());
-    }
-    public List<SimpleFeatureType> createFeatureTypes(WaitPageStatus status) throws Exception{
-        status.setCurrentAction("Databaseverbinding maken...");
-        List<SimpleFeatureType> createdFeatureTypes = new ArrayList<SimpleFeatureType>();
-        DataStore store = null;
-        try {
-            store = createDataStore();
-            status.setProgress(10);
-            status.setCurrentAction("Lijst van tabellen met geo-informatie ophalen...");
-            String[] typeNames = store.getTypeNames();
-            status.setProgress(20);
-
-            if(typeNames.length != 0) {
-                double progress = 20.0;
-                double progressPerTypeName = (80.0/typeNames.length);
-                for(String typeName: typeNames) {
-                    status.setCurrentAction("Inladen schema van tabel \"" + typeName + "\"...");
-                    log.debug("Loading feature source " + typeName + " for JDBCFeatureSource " + getName());
-
-                    SimpleFeatureSource gtFs = store.getFeatureSource(typeName);
-
-                    SimpleFeatureType sft = new SimpleFeatureType();
-                    sft.setTypeName(typeName);
-                    sft.setFeatureSource(this);
-                    sft.setWriteable(true);
-                    if(gtFs.getInfo() != null) {
-                        sft.setDescription(gtFs.getInfo().getDescription());
-                    }
-
-                    org.opengis.feature.simple.SimpleFeatureType gtFt = gtFs.getSchema();
-
-                    for(org.opengis.feature.type.AttributeDescriptor gtAtt: gtFt.getAttributeDescriptors()) {
-                        AttributeDescriptor att = new AttributeDescriptor();
-                        sft.getAttributes().add(att);
-                        att.setName(gtAtt.getLocalName());
-
-                        AttributeType gtType = gtAtt.getType();
-                        String binding = gtType.getBinding().getName();
-
-                        /* XXX use instanceof... */
-                        String type = "";
-                        if(binding.equals("org.locationtech.jts.geom.MultiPolygon")){
-                            type = AttributeDescriptor.TYPE_GEOMETRY_MPOLYGON;
-                        }else if(binding.equals("org.locationtech.jts.geom.Polygon")){
-                            type = AttributeDescriptor.TYPE_GEOMETRY_POLYGON;
-                        }else if(binding.equals("org.locationtech.jts.geom.Geometry")){
-                            type = AttributeDescriptor.TYPE_GEOMETRY;
-                        }else if(binding.equals("org.locationtech.jts.geom.LineString")){
-                            type = AttributeDescriptor.TYPE_GEOMETRY_LINESTRING;
-                        }else if(binding.equals("org.locationtech.jts.geom.Point")){
-                            type = AttributeDescriptor.TYPE_GEOMETRY_POINT;
-                        }else if(binding.equals("org.locationtech.jts.geom.MultiLineString")){
-                            type = AttributeDescriptor.TYPE_GEOMETRY_MLINESTRING;
-                        }else if(binding.equals("org.locationtech.jts.geom.MultiPoint")){
-                            type = AttributeDescriptor.TYPE_GEOMETRY_MPOINT;
-                        }else if(binding.equals("java.lang.Boolean")){
-                            type = AttributeDescriptor.TYPE_BOOLEAN;
-                        }else if(binding.equals("java.lang.Long")){
-                            type = AttributeDescriptor.TYPE_INTEGER;
-                        }else if(binding.equals("java.lang.String")){
-                            type = AttributeDescriptor.TYPE_STRING;
-                        }else if(binding.equals("java.lang.Integer")){
-                            type = AttributeDescriptor.TYPE_INTEGER;
-                        }else if(binding.equals("java.lang.Short")){
-                            type = AttributeDescriptor.TYPE_INTEGER;
-                        }else if(binding.equals("java.lang.Double")){
-                            type = AttributeDescriptor.TYPE_DOUBLE;
-                        }else if(binding.equals("java.lang.Float")){
-                            type = AttributeDescriptor.TYPE_DOUBLE;
-                        }else if(binding.equals("java.sql.Timestamp")){
-                            type = AttributeDescriptor.TYPE_TIMESTAMP;
-                        }else if(binding.equals("java.sql.Date")){
-                            type = AttributeDescriptor.TYPE_DATE;
-                        }else if(binding.equals("java.math.BigDecimal")){
-                            type = AttributeDescriptor.TYPE_DOUBLE;
-                        }
-
-                        if(sft.getGeometryAttribute() == null && gtType instanceof GeometryType) {
-                            sft.setGeometryAttribute(att.getName());
-                        }
-
-                        att.setType(type);
-                    }
-                    createdFeatureTypes.add(sft);
-
-                    progress += progressPerTypeName;
-                    status.setProgress((int)progress);
-                }
-            }
-        } finally {
-            status.setProgress(100);
-            status.setCurrentAction("Databasegegevens ingeladen");
-            status.setFinished(true);
-            if(store != null) {
-                store.dispose();
-            }
-        }
-        return createdFeatureTypes;
-    }
-
-    public DataStore createDataStore() throws Exception {
-        Map params = new HashMap();
-        JSONObject urlObj = new JSONObject(getUrl());
-        params.put("dbtype", urlObj.get("dbtype"));
-        params.put("host", urlObj.get("host"));
-        params.put("port", urlObj.get("port"));
-        params.put("database", urlObj.get("database"));
-
-        params.put("schema", schema);
-        params.put("user", getUsername());
-        params.put(JDBCDataStoreFactory.FETCHSIZE.key,50);
-        params.put("passwd", getPassword());
-        params.put(JDBCDataStoreFactory.EXPOSE_PK.key, true);
-        params.put(JDBCDataStoreFactory.PK_METADATA_TABLE.key, "gt_pk_metadata");
-        // this key is available in ao. Oracle and MS SQL datastore factories, but not in the common parent..
-        // we need this for mssql to determine a featuretype on an empty table
-        if(!urlObj.get("dbtype").equals("oracle")){
-            params.put(SQLServerDataStoreFactory.GEOMETRY_METADATA_TABLE.key, "geometry_columns");
-        }
-        Map logParams = new HashMap(params);
-        if(getPassword() != null) {
-            logParams.put("passwd", new String(new char[getPassword().length()]).replace("\0", "*"));
-        }
-        log.debug("Opening datastore using parameters: " + logParams);
-        try {
-            DataStore ds = DataStoreFinder.getDataStore(params);
-            if(ds == null) {
-                throw new Exception("Cannot open datastore using parameters " + logParams);
-            }
-            return ds;
-        } catch(Exception e) {
-            throw new Exception("Cannot open datastore using parameters " + logParams, e);
-        }
-    }
-
-
-    @Override
-    org.geotools.data.FeatureSource openGeoToolsFeatureSource(SimpleFeatureType sft) throws Exception {
-        DataStore ds = createDataStore();
-
-        return ds.getFeatureSource(sft.getTypeName());
-    }
-
-    @Override
-    org.geotools.data.FeatureSource openGeoToolsFeatureSource(SimpleFeatureType sft, int timeout) throws Exception {
-        return openGeoToolsFeatureSource(sft);
-    }
-
-    @Override
-    FeatureCollection getFeatures(SimpleFeatureType sft, Filter f, int maxFeatures) throws Exception {
-        FeatureCollection fc = null;
-        if(f != null){
-            fc = sft.openGeoToolsFeatureSource().getFeatures(f);
-        }else{
-            fc = sft.openGeoToolsFeatureSource().getFeatures();
-        }
-        return fc;
     }
 }
