+++ viewer-config-persistence/src/main/java/nl/b3p/viewer/config/services/UpdatableFeatureSource.java
@@ -30,90 +30,8 @@ import javax.persistence.EntityManager;
  *
  * @author Roy Braam
  */
-public abstract class UpdatableFeatureSource extends FeatureSource{    
-    private static final Log log = LogFactory.getLog(UpdatableFeatureSource.class);
-    private final int updatebatchsize = 50;
+public abstract class UpdatableFeatureSource extends FeatureSource{
 
-    /**
-     * Update this featuresource.
-     *
-     * @param em the entity manager to use
-     * @return the result of the update
-     * @throws java.lang.Exception if any
-     */
-  /* TODO  public FeatureSourceUpdateResult update(EntityManager em) throws Exception{
-        final FeatureSourceUpdateResult result = new FeatureSourceUpdateResult(this);         
-        try{
-            List<SimpleFeatureType> newFeatureTypes = this.createFeatureTypes(result.getWaitPageStatus().subtask("",80));
-            int processed = 0;
-            //update and add the new featuretypes.
-            for(SimpleFeatureType newFt : newFeatureTypes){
-                MutableBoolean updated = new MutableBoolean();
-                this.addOrUpdateFeatureType(newFt.getTypeName(), newFt, updated);
-
-                MutablePair<SimpleFeatureType,UpdateResult.Status> ftResult = result.getFeatureTypeStatus().get(newFt.getTypeName());
-
-                if (ftResult==null){
-                    result.getFeatureTypeStatus().put(newFt.getTypeName(),new MutablePair(newFt, UpdateResult.Status.NEW));
-                }else{
-                    if(updated.isTrue()) {
-                        log.info("Feature type: "+newFt.getTypeName()+" updated");  
-                        ftResult.setRight(UpdateResult.Status.UPDATED);
-                    }else{
-                        ftResult.setRight(UpdateResult.Status.UNMODIFIED);
-                    }
-                }
-                processed++;
-                if(processed == updatebatchsize){
-                    processed = 0;
-                    if(!em.getTransaction().isActive()){
-                        em.getTransaction().begin();
-                    }
-                    em.persist(this);
-                    em.getTransaction().commit();
-                    em.getTransaction().begin();
-                }
-            }
-            if(!em.getTransaction().isActive()){
-                em.getTransaction().begin();
-            }
-            em.persist(this);
-            em.getTransaction().commit();
-            em.getTransaction().begin();
-            processed = 0;
-            //remove featuretypes when not there
-            Iterator<SimpleFeatureType> it = this.getFeatureTypes().iterator();
-            while (it.hasNext()){
-                SimpleFeatureType oldFt = it.next();
-                boolean stillExists=false;
-                for(SimpleFeatureType newFt : newFeatureTypes){
-                    if (newFt.getTypeName().equals(oldFt.getTypeName())){
-                        stillExists=true;
-                        break;
-                    }
-                }
-                if(!stillExists){
-                    it.remove();
-                    em.remove(oldFt);
-                }
-                if(processed == updatebatchsize){
-                    processed = 0;
-                    if(!em.getTransaction().isActive()){
-                        em.getTransaction().begin();
-                    }
-                    em.persist(this);
-                    em.getTransaction().commit();
-                    em.getTransaction().begin();
-                }
-            }
-            result.setStatus(UpdateResult.Status.UPDATED);
-            
-        }catch(Exception e){
-            result.failedWithException(e);        
-        }
-        return result;
-    }*/
-    
     /**
      * return a list of featuretypes that are currently present in the
      * FeatureSource.
