+++ viewer/src/main/java/nl/b3p/viewer/stripes/DownloadFeaturesActionBean.java
@@ -55,6 +55,8 @@ import nl.b3p.viewer.features.ExcelDownloader;
 import nl.b3p.viewer.features.FeatureDownloader;
 import nl.b3p.viewer.features.GeoJSONDownloader;
 import nl.b3p.viewer.features.ShapeDownloader;
+import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
+import nl.b3p.viewer.helpers.featuresources.WFSSourceHelper;
 import nl.b3p.viewer.util.ChangeMatchCase;
 import nl.b3p.viewer.util.FeatureToJson;
 import nl.b3p.viewer.util.FlamingoCQL;
@@ -243,10 +245,10 @@ public class DownloadFeaturesActionBean extends LocalizableApplicationActionBean
                 if (isDebug() && ft.getFeatureSource() instanceof WFSFeatureSource) {
                     Map extraDataStoreParams = new HashMap();
                     extraDataStoreParams.put(WFSDataStoreFactory.TRY_GZIP.key, Boolean.FALSE);
-                    fs = ((WFSFeatureSource) ft.getFeatureSource()).openGeoToolsFeatureSource(layer.getFeatureType(), extraDataStoreParams);
+                    SimpleFeatureType sft = layer.getFeatureType();
+                    fs = WFSSourceHelper.openGeoToolsFSFeatureSource(sft, extraDataStoreParams, (WFSFeatureSource) sft.getFeatureSource());
                 } else {
-
-                    fs = ft.openGeoToolsFeatureSource();
+                    fs = SourceFactoryHelper.openGeoToolsFeatureSource(ft.getFeatureSource(), ft);
                 }
 
                 final Query q = new Query(fs.getName().toString());
