+++ viewer/src/main/java/nl/b3p/viewer/stripes/EditFeatureActionBean.java
@@ -17,6 +17,7 @@
 package nl.b3p.viewer.stripes;
 
 import net.sourceforge.stripes.controller.LifecycleStage;
+import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.io.WKTReader;
 import java.io.IOException;
@@ -189,7 +190,7 @@ public class EditFeatureActionBean extends LocalizableApplicationActionBean impl
                     break;
                 }
 
-                fs = layer.getFeatureType().openGeoToolsFeatureSource();
+                fs = SourceFactoryHelper.openGeoToolsFeatureSource(layer.getFeatureType());
 
                 if(!(fs instanceof SimpleFeatureStore)) {
                     error = getBundle().getString("viewer.editfeatureactionbean.5");
@@ -269,7 +270,7 @@ public class EditFeatureActionBean extends LocalizableApplicationActionBean impl
                         SimpleFeatureType fType = rel.getForeignFeatureType();
                         label = fType.getDescription() == null ? fType.getTypeName() : fType.getDescription();
 
-                        fs = fType.openGeoToolsFeatureSource(5000);
+                        fs = SourceFactoryHelper.openGeoToolsFeatureSource(fType,5000);
                         store = (SimpleFeatureStore) fs;
                         jsonFeature = new JSONObject(feature);
                         String fid = jsonFeature.optString(FID, null);
@@ -373,7 +374,7 @@ public class EditFeatureActionBean extends LocalizableApplicationActionBean impl
                     break;
                 }
 
-                fs = layer.getFeatureType().openGeoToolsFeatureSource();
+                fs = SourceFactoryHelper.openGeoToolsFeatureSource(layer.getFeatureType());
 
                 if(!(fs instanceof SimpleFeatureStore)) {
                     error = getBundle().getString("viewer.editfeatureactionbean.13");
@@ -444,7 +445,7 @@ public class EditFeatureActionBean extends LocalizableApplicationActionBean impl
                 if (rel.getType().equals(FeatureTypeRelation.RELATE)) {
                     SimpleFeatureType fType = rel.getForeignFeatureType();
                     label = fType.getDescription() == null ? fType.getTypeName() : fType.getDescription();
-                    fs = fType.openGeoToolsFeatureSource(5000);
+                    fs = SourceFactoryHelper.openGeoToolsFeatureSource(fType,5000);
                     store = (SimpleFeatureStore) fs;
                     jsonFeature = new JSONObject(feature);
                     String fid = jsonFeature.optString(FID, null);
