+++ viewer/src/main/java/nl/b3p/viewer/stripes/FeatureReportActionBean.java
@@ -16,6 +16,7 @@
  */
 package nl.b3p.viewer.stripes;
 
+import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.simplify.TopologyPreservingSimplifier;
 import net.sourceforge.stripes.action.*;
@@ -137,7 +138,7 @@ public class FeatureReportActionBean extends LocalizableApplicationActionBean im
             JSONObject params = new JSONObject(printparams);
 
             FilterFactory2 ff = CommonFactoryFinder.getFilterFactory2();
-            FeatureSource fs = layer.getFeatureType().openGeoToolsFeatureSource(TIMEOUT);
+            FeatureSource fs = SourceFactoryHelper.openGeoToolsFeatureSource(layer.getFeatureType(), TIMEOUT);
             String geomAttribute = fs.getSchema().getGeometryDescriptor().getLocalName();
             Filter fidFilter = ff.id(new FeatureIdImpl(this.fid));
 
@@ -240,7 +241,7 @@ public class FeatureReportActionBean extends LocalizableApplicationActionBean im
                             q.setHandle("FeatureReportActionBean_related_attributes");
                             LOG.debug("Related features query: " + q);
 
-                            fs = fType.openGeoToolsFeatureSource(TIMEOUT);
+                            fs = SourceFactoryHelper.openGeoToolsFeatureSource(fType, TIMEOUT);
                             features = ftjson.getJSONFeatures(appLayer, fType, fs, q, em, application, context.getRequest());
 
                             JSONArray jsonFeats = new JSONArray();
