+++ viewer/src/main/java/nl/b3p/viewer/stripes/ServiceActionBean.java
@@ -28,8 +28,8 @@ import net.sourceforge.stripes.action.StreamingResolution;
 import net.sourceforge.stripes.action.StrictBinding;
 import net.sourceforge.stripes.action.UrlBinding;
 import net.sourceforge.stripes.validation.Validate;
-import nl.b3p.viewer.ArcGISServiceHelper;
-import nl.b3p.viewer.WMSServiceHelper;
+import nl.b3p.viewer.helpers.services.ArcGISServiceHelper;
+import nl.b3p.viewer.helpers.services.WMSServiceHelper;
 import nl.b3p.viewer.config.services.ArcGISService;
 import nl.b3p.viewer.config.services.GeoService;
 import nl.b3p.viewer.config.services.WMSService;
