+++ viewer/src/main/java/nl/b3p/viewer/userlayer/TailormapDBManager.java
@@ -1,7 +1,10 @@
 package nl.b3p.viewer.userlayer;
 
-import nl.b3p.viewer.GeoserviceFactoryHelper;
-import nl.b3p.viewer.WMSServiceHelper;
+import nl.b3p.viewer.helpers.featuresources.FeatureSourceHelper;
+import nl.b3p.viewer.helpers.featuresources.JDBCSourceHelper;
+import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
+import nl.b3p.viewer.helpers.services.GeoserviceFactoryHelper;
+import nl.b3p.viewer.helpers.services.WMSServiceHelper;
 import nl.b3p.viewer.audit.AuditMessageObject;
 import nl.b3p.viewer.config.ClobElement;
 import nl.b3p.viewer.config.app.Application;
@@ -136,7 +139,7 @@ public class TailormapDBManager {
             JDBCFeatureSource fs = (JDBCFeatureSource)this.layer.getFeatureType().getFeatureSource();
             try {
                 Date start = new Date();
-                // TODO fs.update(entityManager);
+                FeatureSourceHelper.update(entityManager,fs);
                 Date end = new Date();
                 LOG.error("time: " + (end.getTime() - start.getTime()));
                 SimpleFeatureType newFt = fs.getFeatureType(viewName);
