+++ viewer/src/main/java/nl/b3p/viewer/userlayer/UserLayerHandler.java
@@ -1,7 +1,9 @@
 package nl.b3p.viewer.userlayer;
 
 
-import nl.b3p.viewer.WMSServiceHelper;
+import nl.b3p.viewer.helpers.featuresources.JDBCSourceHelper;
+import nl.b3p.viewer.helpers.featuresources.SourceFactoryHelper;
+import nl.b3p.viewer.helpers.services.WMSServiceHelper;
 import nl.b3p.viewer.audit.AuditMessageObject;
 import nl.b3p.viewer.config.app.Application;
 import nl.b3p.viewer.config.app.ApplicationLayer;
@@ -10,10 +12,8 @@ import nl.b3p.viewer.util.FlamingoCQL;
 import nl.b3p.web.WaitPageStatus;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.geotools.data.jdbc.FilterToSQL;
 import org.geotools.data.jdbc.FilterToSQLException;
 import org.geotools.filter.text.cql2.CQLException;
-import org.geotools.jdbc.BasicSQLDialect;
 import org.geotools.jdbc.JDBCDataStore;
 
 import javax.persistence.EntityManager;
@@ -70,7 +70,7 @@ public class UserLayerHandler {
             this.service = this.appLayer.getService();
             this.layer = this.service.getLayer(this.appLayer.getLayerName(), this.entityManager);
             this.tableName = this.layer.getFeatureType().getTypeName();
-            this.dataStore = (JDBCDataStore) layer.getFeatureType().openGeoToolsFeatureSource().getDataStore();
+            this.dataStore = (JDBCDataStore) SourceFactoryHelper.openGeoToolsFeatureSource(layer.getFeatureType()).getDataStore();
             this.dataBase = DataBaseFactory.getDataBase(dataStore);
         } catch (Exception e) {
             LOG.fatal("Problem opening datastore. " + e.getLocalizedMessage());
